requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.14
django-celery-results==2.4.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
